# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "这个API接口文档",
        "title": "LeeGo",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "寒云",
            "url": "http://www.swagger.io/support",
            "email": "1355081829@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/admin/v1/admin/list": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "description": "# 请求参数\n| 参数 | 说明 | 备注 |\n| :-----| :----- | :----- |\n|page|页码|默认是1，可以不填|\n# 返回参数参数\n| 参数 | 说明 | 备注 |\n| :-----| :----- | :----- |\n|name|用户名||\n|avatar|头像||\n|roleId|角色id||\n|limit|每页数量||\n|totalCount|总数据量||\n|totalPage|总页码数||\n|status|状态|20000表示成功|\n|code|状态码||\n|message|返回的信息||\n|data|返回的具体数据||",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/v1/login": {
            "post": {
                "description": "# 请求参数\n| 参数 | 说明 | 备注 |\n| :-----| :----- | :----- |\n|username|用户名||\n|password|密码||\n# 返回参数参数\n| 参数 | 说明 | 备注 |\n| :-----| :----- | :----- |\n|code|状态码||\n|message|返回的信息||\n|data|返回的具体数据||",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/v1/logout": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "退出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/v1/order/list": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "description": "# 请求参数\n| 参数 | 说明 | 备注 |\n| :-----| :----- | :----- |\n|page|页码||\n# 返回参数参数\n| 参数 | 说明 | 备注 |\n| :-----| :----- | :----- |\n|code|状态码||\n|message|返回的信息||\n|data|返回的具体数据||",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "订单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/v1/user/list": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "description": "# 请求参数\n| 参数 | 说明 | 备注 |\n| :-----| :----- | :----- |\n|page|页码||\n# 返回参数参数\n| 参数 | 说明 | 备注 |\n| :-----| :----- | :----- |\n|code|状态码||\n|message|返回的信息||\n|data|返回的具体数据||",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/": {
            "get": {
                "description": "# aaaaa",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "首页"
                ],
                "summary": "首页",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "ID",
                        "name": "img",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/index/detail": {
            "get": {
                "description": "id 文章id|name文章名称",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "首页"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/index/post": {
            "post": {
                "description": "id 文章id|name文章名称",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "首页"
                ],
                "summary": "详情",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "img",
                        "name": "img",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}